+++ snprc_ehr/src/org/labkey/snprc_ehr/model/CalculatedColumnQueryInfo.java
@@ -1,4 +1,4 @@
-package org.labkey.snprc_ehr.domain;
+package org.labkey.snprc_ehr.model;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -6,7 +6,7 @@ import lombok.Setter;
 import org.labkey.api.data.AbstractTableInfo;
 import org.labkey.api.data.ColumnInfo;
 import org.labkey.api.query.UserSchema;
-import org.labkey.snprc_ehr.domain.CalculatedColumn;
+import org.labkey.snprc_ehr.model.CalculatedColumn;
 
 import java.util.List;
